+++ pyscriptjs/src/pyconfig.ts
@@ -54,8 +54,8 @@ export const defaultConfig: AppConfig = {
     type: 'app',
     interpreters: [
         {
-            src: 'https://cdn.jsdelivr.net/pyodide/v0.22.1/full/pyodide.js',
-            name: 'pyodide-0.22.1',
+            src: 'https://cdn.jsdelivr.net/pyodide/v0.23.2/full/pyodide.js',
+            name: 'pyodide-0.23.2',
             lang: 'python',
         },
     ],