+++ pyscriptjs/tests/integration/test_zz_examples.py
@@ -276,18 +276,20 @@ class TestExamples(PyScriptTest):
         self.check_tutor_generated_code()
 
     def test_panel_kmeans(self):
-        # XXX improve this test
+        # XXX improve this test>>>>>>> main
         self.goto("examples/panel_kmeans.html")
-        self.wait_for_pyscript(timeout=90 * 1000)
+        self.wait_for_pyscript(timeout=120 * 1000)
         assert self.page.title() == "Pyscript/Panel KMeans Demo"
-        wait_for_render(self.page, "*", "<div.*?class=['\"]bk-root['\"].*?>")
+        wait_for_render(
+            self.page, "*", "<div.*?class=['\"]bk-root['\"].*?>", timeout_seconds=60 * 2
+        )
         self.assert_no_banners()
         self.check_tutor_generated_code()
 
     def test_panel_stream(self):
         # XXX improve this test
         self.goto("examples/panel_stream.html")
-        self.wait_for_pyscript(timeout=90 * 1000)
+        self.wait_for_pyscript(timeout=3 * 60 * 1000)
         assert self.page.title() == "PyScript/Panel Streaming Demo"
         wait_for_render(self.page, "*", "<div.*?class=['\"]bk-root['\"].*?>")
         self.assert_no_banners()
@@ -316,7 +318,7 @@ class TestExamples(PyScriptTest):
 
     def test_repl2(self):
         self.goto("examples/repl2.html")
-        self.wait_for_pyscript()
+        self.wait_for_pyscript(timeout=1.5 * 60 * 1000)
         assert self.page.title() == "Custom REPL Example"
         wait_for_render(self.page, "*", "<py-repl.*?>")
         # confirm we can import utils and run one command